requirements.txt
requests
python-dotenv
retry
kafka-python
